dataprocpy/requirements.txt
requests==2.20.0
google-cloud-dataproc
flask
flask-cors